+++ web_server_with_database/src/orca_slicer_interface/mod.rs
@@ -1,2 +1,2 @@
-mod orca_slicer_interface_utilities;
-pub use orca_slicer_interface_utilities::initialize_orca_slicer_if;
+mod orca_slicer_interface;
+pub use orca_slicer_interface::initialize_orca_slicer_if;
